refactor: api.service 디렉토리 수정
