+++ src/api/services/fetch/fetchService.test.ts
@@ -2,7 +2,7 @@ import { beforeEach, expect, test, vi } from "vitest";
 import { mock } from "vitest-mock-extended";
 import { createGitHubClient } from "../../infra/gitHub/gitHubClient";
 import { createFetchService } from "./fetchService";
-import { Grade } from "../common/types";
+import { Grade } from "../types";
 import { GitHubMember, GitHubTeam, GitHubTree } from "../../infra/gitHub/types";
 
 // Mock data
